Test moving transition and doCalcdata to plots.js
